move geobuf and indexing tests to tape
